# Use a stack-allocated array for script counters