Fix the todo in detect_script.